SobolGenerator now returns a vector from the unit hypercube. Monte Carlo and timed generation also added. Current time for 20000000 iterations of 4 dimensional vectors is 2.563.

+++ SobolGenerator.hpp
@@ -32,7 +32,7 @@ public:
     // destructor
     ~SobolGenerator();
     friend ostream& operator<<(ostream&, const SobolGenerator&);
-    vector<unsigned> get_next();
+    vector<double> get_next();
 };
